[monitoring/rid_qualifier] Rename host in webapp
