+++ monitoring/rid_qualifier/webapp/__init__.py
@@ -6,4 +6,4 @@ webapp = flask.Flask(__name__)
 
 webapp.config.from_object(Config)
 
-from monitoring.rid_qualifier.host import routes
+from monitoring.rid_qualifier.webapp import routes
