+++ monitoring/rid_qualifier/webapp/debug_host.py
@@ -1,6 +1,6 @@
 import sys
 
-from monitoring.rid_qualifier.host import webapp
+from monitoring.rid_qualifier.webapp import webapp
 from . import config
